Major dmnsn_kD_splay_search() optimization.
At each level of recursion, we have to go down the right branch if it exists.
But if we do this before we test the current node and the left branch, we can
eliminate those tests in the likely case that we find a closer object in the
geometrically larger right subtree.  This gives about a 2X speed improvement
according to `make bench'.